requirements.txt
flask==2.2.2
flask-cors==3.0.10
torch==2.0.1
torchvision==0.15.2
Pillow==9.4.0
transformers==4.30.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.20.9.\venv\Scripts\activate
Werkzeug==2.3.8